Refactor:
- Changed url parameters to display vendor according to params recieved
- Removed pagination and filter component and added it into vendorLayout
