+++ src/components/constants.js
@@ -1,0 +1,1406 @@
+export const karachiAreas=[
+  {
+   
+    "name":"Askari 1"
+  },
+  {
+   
+    "name":"Askari 2"
+  },
+  {
+  
+    "name":"Askari 3"
+  },
+  {
+    
+    "name":"Askari 4"
+  },
+  {
+    
+    "name":"Askari 5"
+  },
+  {
+    
+    "name":"Bahria Town - Precinct 1"
+  },
+  {
+    "name":"Bahria Town - Precinct 10"
+  },
+  {
+    "name":"Bahria Town - Precinct 11"
+  },
+  {
+    "name":"Bahria Town - Precinct 12"
+  },
+  {
+    "name":"Bahria Town - Precinct 13"
+  },
+  {
+    "name":"Bahria Town - Precinct 14"
+  },
+  {
+    "name":"Bahria Town - Precinct 15"
+  },
+  {
+    "name":"Bahria Town - Precinct 16"
+  },
+  {
+    "name":"Bahria Town - Precinct 17"
+  },
+  {
+    "name":"Bahria Town - Precinct 18"
+  },
+  {
+    "name":"Bahria Town - Precinct 19"
+  },
+  {
+    "name":"Bahria Town - Precinct 2"
+  },
+  {
+    "name":"Bahria Town - Precinct 20"
+  },
+  {
+    "name":"Bahria Town - Precinct 21"
+  },
+  {
+    "name":"Bahria Town - Precinct 22"
+  },
+  {
+    "name":"Bahria Town - Precinct 23"
+  },
+  {
+    "name":"Bahria Town - Precinct 24"
+  },
+  {
+    "name":"Bahria Town - Precinct 25"
+  },
+  {
+    "name":"Bahria Town - Precinct 26"
+  },
+  {
+    "name":"Bahria Town - Precinct 27"
+  },
+  {
+    "name":"Bahria Town - Precinct 28"
+  },
+  {
+    "name":"Bahria Town - Precinct 29"
+  },
+  {
+    "name":"Bahria Town - Precinct 3"
+  },
+  {
+    "name":"Bahria Town - Precinct 30"
+  },
+  {
+    "name":"Bahria Town - Precinct 31"
+  },
+  {
+    "name":"Bahria Town - Precinct 32"
+  },
+  {
+    "name":"Bahria Town - Precinct 33"
+  },
+  {
+    "name":"Bahria Town - Precinct 4"
+  },
+  {
+    "name":"Bahria Town - Precinct 5"
+  },
+  {
+    "name":"Bahria Town - Precinct 6"
+  },
+  {
+    "name":"Bahria Town - Precinct 7"
+  },
+  {
+    "name":"Bahria Town - Precinct 8"
+  },
+  {
+    "name":"Bahria Town - Precinct 9"
+  },
+  {
+    "name":"BufferZone - Sector 15 A 1"
+  },
+  {
+    "name":"BufferZone - Sector 15 A 2"
+  },
+  {
+    "name":"BufferZone - Sector 15 A 3"
+  },
+  {
+    "name":"BufferZone - Sector 15 A 4"
+  },
+  {
+    "name":"BufferZone - Sector 15 A 5"
+  },
+  {
+    "name":"BufferZone - Sector 15 B"
+  },
+  {
+    "name":"BufferZone - Sector 16 A"
+  },
+  {
+    "name":"BufferZone - Sector 16 B"
+  },
+  {
+    "name":"Cantonment"
+  },
+  {
+    "name":"Clifton - Block 1"
+  },
+  {
+    "name":"Clifton - Block 2"
+  },
+  {
+    "name":"Clifton - Block 3"
+  },
+  {
+    "name":"Clifton - Block 4"
+  },
+  {
+    "name":"Clifton - Block 5"
+  },
+  {
+    "name":"Clifton - Block 6"
+  },
+  {
+    "name":"Clifton - Block 7"
+  },
+  {
+    "name":"Clifton - Block 8"
+  },
+  {
+    "name":"Clifton - Block 9"
+  },
+  {
+    "name":"Clifton - Kehkashan"
+  },
+  {
+    "name":"DHA - Phase 1"
+  },
+  {
+    "name":"DHA - Phase 2"
+  },
+  {
+    "name":"DHA - Phase 3"
+  },
+  {
+    "name":"DHA - Phase 4"
+  },
+  {
+    "name":"DHA - Phase 5"
+  },
+  {
+    "name":"DHA - Phase 6"
+  },
+  {
+    "name":"DHA - Phase 7"
+  },
+  {
+    "name":"DHA - Phase 8"
+  },
+  {
+    "name":"DHA - Phase 9"
+  },
+  {
+    "name":"F.B Area - Azizabad"
+  },
+  {
+    "name":"F.B Area - B1 Area"
+  },
+  {
+    "name":"F.B Area - B Area"
+  },
+  {
+    "name":"F.B Area - Block 1"
+  },
+  {
+    "name":"F.B Area - Block 10"
+  },
+  {
+    "name":"F.B Area - Block 11"
+  },
+  {
+    "name":"F.B Area - Block 12"
+  },
+  {
+    "name":"F.B Area - Block 13"
+  },
+  {
+    "name":"F.B Area - Block 14"
+  },
+  {
+    "name":"F.B Area - Block 15"
+  },
+  {
+    "name":"F.B Area - Block 16"
+  },
+  {
+    "name":"F.B Area - Block 17"
+  },
+  {
+    "name":"F.B Area - Block 18"
+  },
+  {
+    "name":"F.B Area - Block 19"
+  },
+  {
+    "name":"F.B Area - Block 2"
+  },
+  {
+    "name":"F.B Area - Block 20"
+  },
+  {
+    "name":"F.B Area - Block 21"
+  },
+  {
+    "name":"F.B Area - Block 22"
+  },
+  {
+    "name":"F.B Area - Block 3"
+  },
+  {
+    "name":"F.B Area - Block 4"
+  },
+  {
+    "name":"F.B Area - Block 5"
+  },
+  {
+    "name":"F.B Area - Block 6"
+  },
+  {
+    "name":"F.C Area - C1 Area"
+  },
+  {
+    "name":"F.C Area - C Area"
+  },
+  {
+    "name":"Garden - Garden East"
+  },
+  {
+    "name":"Garden - Garden West"
+  },
+  {
+    "name":"Garden - Soldier Bazaar"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 1"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 10"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 11"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 12"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 13"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 14"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 15"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 16"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 17"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 18"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 19"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 2"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 20"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 3"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 4"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 5"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 6"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 7"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 8"
+  },
+  {
+    "name":"Gulistan-e-Johar - Block 9"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Data Nagar"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - EIDU Goth"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Gulshan-e-Mauzzam"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Gulshan-e-Rehman"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Mehran Road"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Phase 1"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Phase 2"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Phase 3"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - PTCL Satellite Station"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Shah Latif Town"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Shahnawaz Goth"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Shah Town"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - Steel Town"
+  },
+  {
+    "name":"Gulshan-e-Hadeed - TCP Godowns"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Adamjee Nagar"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 1"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 10"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 11"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 12"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 13"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 14"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 15"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 16"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 17"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 18"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 19"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 2"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 3"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 4"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 5"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 6"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 7"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 8"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Block 9"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Civic Center"
+  },
+  {
+    "name":"Gulshan-e-Iqbal - Dhoraji"
+  },
+  {
+    "name":"Korangi - Abdullah Shah Noorani Pahari Colony"
+  },
+  {
+    "name":"Korangi - Korangi Industrial Area"
+  },
+  {
+    "name":"Korangi - Nasir Colony"
+  },
+  {
+    "name":"Korangi - PAF Base Korangi Creek"
+  },
+  {
+    "name":"Korangi - Zaman Town"
+  },
+  {
+    "name":"Korangi - Zia Colony"
+  },
+  {
+    "name":"Landhi - Alflah Housing Society"
+  },
+  {
+    "name":"Landhi - Awami Colony"
+  },
+  {
+    "name":"Landhi - Bagh-e-Korangi"
+  },
+  {
+    "name":"Landhi - Bakhtawar Goth"
+  },
+  {
+    "name":"Landhi - Barmi Colony"
+  },
+  {
+    "name":"Landhi - Bhutto Nagar"
+  },
+  {
+    "name":"Landhi - Future Colony"
+  },
+  {
+    "name":"Landhi - Gulshan-e-Rafi"
+  },
+  {
+    "name":"Landhi - Ilyas Goth"
+  },
+  {
+    "name":"Landhi - Labour Colony"
+  },
+  {
+    "name":"Landhi - Landhi Industrial Area"
+  },
+  {
+    "name":"Landhi - Muslimabad Colony"
+  },
+  {
+    "name":"Landhi - Muzaffarabad Colony"
+  },
+  {
+    "name":"Landhi - Punjab Town"
+  },
+  {
+    "name":"Landhi - Qasim Town"
+  },
+  {
+    "name":"Landhi - Sadat Colony"
+  },
+  {
+    "name":"Landhi - Shah Khalid Colony"
+  },
+  {
+    "name":"Landhi - Sharafi Goth"
+  },
+  {
+    "name":"Landhi - Zamanabad"
+  },
+  {
+    "name":"Liaquatabad - Block 1"
+  },
+  {
+    "name":"Liaquatabad - Block 10"
+  },
+  {
+    "name":"Liaquatabad - Block 2"
+  },
+  {
+    "name":"Liaquatabad - Block 3"
+  },
+  {
+    "name":"Liaquatabad - Block 4"
+  },
+  {
+    "name":"Liaquatabad - Block 5"
+  },
+  {
+    "name":"Liaquatabad - Block 6"
+  },
+  {
+    "name":"Liaquatabad - Block 7"
+  },
+  {
+    "name":"Liaquatabad - Block 8"
+  },
+  {
+    "name":"Liaquatabad - Block 9"
+  },
+  {
+    "name":"Malir - Malir"
+  },
+  {
+    "name":"Malir - Malir Cantt"
+  },
+  {
+    "name":"Nazimabad - Block 1"
+  },
+  {
+    "name":"Nazimabad - Block 2"
+  },
+  {
+    "name":"Nazimabad - Block 3"
+  },
+  {
+    "name":"Nazimabad - Block 4"
+  },
+  {
+    "name":"Nazimabad - Block 5"
+  },
+  {
+    "name":"North Karachi  - Sector 10"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - A"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - B"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - C 1"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - C 2"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - C 3"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - E"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - H"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - I"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - K"
+  },
+  {
+    "name":"North Karachi  - Sector 11 - L"
+  },
+  {
+    "name":"North Karachi  - Sector 2"
+  },
+  {
+    "name":"North Karachi  - Sector 3"
+  },
+  {
+    "name":"North Karachi  - Sector 4"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - A 1"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - A 2"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - A 3"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - A 4"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - B 1"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - B 2"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - B 3"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - B 4"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - C 1"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - C 2"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - C 3"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - C 4"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - I"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - J"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - K"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - L"
+  },
+  {
+    "name":"North Karachi  - Sector 5 - M"
+  },
+  {
+    "name":"North Karachi  - Sector 6"
+  },
+  {
+    "name":"North Karachi  - Sector 7 - D 1"
+  },
+  {
+    "name":"North Karachi  - Sector 7 - D 2"
+  },
+  {
+    "name":"North Karachi  - Sector 7 - D 3"
+  },
+  {
+    "name":"North Karachi  - Sector 7 - D 4"
+  },
+  {
+    "name":"North Karachi  - Sector 8"
+  },
+  {
+    "name":"North Karachi  - Sector 9"
+  },
+  {
+    "name":"North Nazimabad - Block A"
+  },
+  {
+    "name":"North Nazimabad - Block B"
+  },
+  {
+    "name":"North Nazimabad - Block C"
+  },
+  {
+    "name":"North Nazimabad - Block D"
+  },
+  {
+    "name":"North Nazimabad - Block E"
+  },
+  {
+    "name":"North Nazimabad - Block F"
+  },
+  {
+    "name":"North Nazimabad - Block G"
+  },
+  {
+    "name":"North Nazimabad - Block H"
+  },
+  {
+    "name":"North Nazimabad - Block I"
+  },
+  {
+    "name":"North Nazimabad - Block J"
+  },
+  {
+    "name":"North Nazimabad - Block K"
+  },
+  {
+    "name":"North Nazimabad - Block L"
+  },
+  {
+    "name":"North Nazimabad - Block M"
+  },
+  {
+    "name":"North Nazimabad - Block N"
+  },
+  {
+    "name":"North Nazimabad - Block O"
+  },
+  {
+    "name":"North Nazimabad - Block P"
+  },
+  {
+    "name":"North Nazimabad - Block Q"
+  },
+  {
+    "name":"North Nazimabad - Block R"
+  },
+  {
+    "name":"North Nazimabad - Block S"
+  },
+  {
+    "name":"North Nazimabad - Block T"
+  },
+  {
+    "name":"Old Town - Bhimpora"
+  },
+  {
+    "name":"Old Town - Bohra Pir"
+  },
+  {
+    "name":"Old Town - Bombay Bazar"
+  },
+  {
+    "name":"Old Town - Jodia Bazar"
+  },
+  {
+    "name":"Old Town - Kagzi Bazar"
+  },
+  {
+    "name":"Old Town - Kakri Ground"
+  },
+  {
+    "name":"Old Town - Kamil Gali"
+  },
+  {
+    "name":"Old Town - Khada Market"
+  },
+  {
+    "name":"Old Town - Kharadar"
+  },
+  {
+    "name":"Old Town - Lee Market"
+  },
+  {
+    "name":"Old Town - Mithadar"
+  },
+  {
+    "name":"Old Town - Nanwara"
+  },
+  {
+    "name":"Old Town - Nishter Road"
+  },
+  {
+    "name":"Old Town - Pan Mandi"
+  },
+  {
+    "name":"Old Town - Ramswami"
+  },
+  {
+    "name":"Old Town - Ranchorline"
+  },
+  {
+    "name":"Orangi Town - Banaras Town"
+  },
+  {
+    "name":"Orangi Town - Bangla Bazaar"
+  },
+  {
+    "name":"Orangi Town - Bilal Colony"
+  },
+  {
+    "name":"Orangi Town - Katti Pahari"
+  },
+  {
+    "name":"Orangi Town - Moria Goth Orangi"
+  },
+  {
+    "name":"Orangi Town - Orangi"
+  },
+  {
+    "name":"Orangi Town - Sector 14 - A"
+  },
+  {
+    "name":"Orangi Town - Sector 14 - C"
+  },
+  {
+    "name":"Orangi Town - Thorani Goth"
+  },
+  {
+    "name":"Baldiya Town"
+  },
+  {
+    "name":"Baloch Colony"
+  },
+  {
+    "name":"Civil Line"
+  },
+  {
+    "name":"FC Area"
+  },
+  {
+    "name":"Firdous Colony"
+  },
+  {
+    "name":"Gulshan-e-Maymar"
+  },
+  {
+    "name":"Hawksbay"
+  },
+  {
+    "name":"I.I Chundrigar"
+  },
+  {
+    "name":"Jamshed Road"
+  },
+  {
+    "name":"K.D.A Officers"
+  },
+  {
+    "name":"Kemari"
+  },
+  {
+    "name":"Liyari"
+  },
+  {
+    "name":"M.A Jinnah Rd"
+  },
+  {
+    "name":"Manora"
+  },
+  {
+    "name":"New Karachi"
+  },
+  {
+    "name":"New Surjani"
+  },
+  {
+    "name":"PIB Colony"
+  },
+  {
+    "name":"Pipri Goth"
+  },
+  {
+    "name":"Rizvia Society"
+  },
+  {
+    "name":"Saddar"
+  },
+  {
+    "name":"Scheme 33"
+  },
+  {
+    "name":"Shabbirabad"
+  },
+  {
+    "name":"P.E.C.H.S - Block 1"
+  },
+  {
+    "name":"P.E.C.H.S - Block 2"
+  },
+  {
+    "name":"P.E.C.H.S - Block 3"
+  },
+  {
+    "name":"P.E.C.H.S - Block 4"
+  },
+  {
+    "name":"P.E.C.H.S - Block 5"
+  },
+  {
+    "name":"P.E.C.H.S - Block 6"
+  },
+  {
+    "name":"P.E.C.H.S - Khalid Bin Walid"
+  },
+  {
+    "name":"P.E.C.H.S - Tariq Road"
+  },
+  {
+    "name":"S.I.T.E - Golimar"
+  },
+  {
+    "name":"S.I.T.E - S.I.T.E"
+  },
+  {
+    "name":"Shah Faisal Colony - Aswan Town"
+  },
+  {
+    "name":"Shah Faisal Colony - Gulshan-e-Asghar"
+  },
+  {
+    "name":"Shah Faisal Colony - Shah Faisal Colony 1"
+  },
+  {
+    "name":"Shah Faisal Colony - Shah Faisal Colony 5"
+  },
+  {
+    "name":"F.B Area – Block 7"
+  },
+  {
+    "name":"F.B Area - Block 9"
+  },
+  {
+    "name":"P.E.C.H.S - Block 7"
+  },
+  {
+    "name":"Aram Bagh"
+  },
+  {
+    "name":"Bath Island"
+  },
+  {
+    "name":"University Road"
+  },
+  {
+    "name":"Bahadurabad"
+  },
+  {
+    "name":"Shah Faisal Colony - 4"
+  },
+  {
+    "name":"Banglore Town"
+  },
+  {
+    "name":"Fowler Lines"
+  },
+  {
+    "name":"Shah Faisal Colony - Shamsi Society"
+  },
+  {
+    "name":"Gulshan-e-Jamal"
+  },
+  {
+    "name":"Shah Faisal Colony - 3"
+  },
+  {
+    "name":"Shah Faisal Colony - Green Town"
+  },
+  {
+    "name":"Darwaish Colony"
+  },
+  {
+    "name":"Korangi - Sector 31 B"
+  },
+  {
+    "name":"Firdous Colony"
+  },
+  {
+    "name":"North Nazimabad - Block W"
+  },
+  {
+    "name":"K.A.E.C.H.S"
+  },
+  {
+    "name":"Mehmoodabad"
+  },
+  {
+    "name":"Korangi - Mehran Town"
+  },
+  {
+    "name":"Landhi Town - 36 B"
+  },
+  {
+    "name":"Karachi Memon Society"
+  },
+  {
+    "name":"Madras Cooperative Housing Society"
+  },
+  {
+    "name":"Shahrah-e-Faisal"
+  },
+  {
+    "name":"Korangi - Sector 41 B"
+  },
+  {
+    "name":"Clifton - Delhi Colony"
+  },
+  {
+    "name":"Korangi - Sector 32 B"
+  },
+  {
+    "name":"Dhoraji - Adamjee Nagar"
+  },
+  {
+    "name":"Bhimpura"
+  },
+  {
+    "name":"Dhoraji - CP & Berar Society"
+  },
+  {
+    "name":"Shahra-e-Faisal - Umar Colony"
+  },
+  {
+    "name":"Model Colony"
+  },
+  {
+    "name":"Gulshan-e-Shamim"
+  },
+  {
+    "name":"Clifton - Shah Rasool Colony"
+  },
+  {
+    "name":"North Karachi  - Sector 12 C"
+  },
+  {
+    "name":"Jail Road - Hyderabad Colony"
+  },
+  {
+    "name":"Napier Quarter"
+  },
+  {
+    "name":"Gulzar-e-Hijri"
+  },
+  {
+    "name":"North Karachi  - Sector 12 A"
+  },
+  {
+    "name":"Shahra-e-Faisal - Jinnah Housing Society"
+  },
+  {
+    "name":"K.D.A Scheme 1"
+  },
+  {
+    "name":"Clifton - Punjab Colony"
+  },
+  {
+    "name":"Korangi - Sector 31 D"
+  },
+  {
+    "name":"Clifton - Zamzama"
+  },
+  {
+    "name":"Parsi Colony"
+  },
+  {
+    "name":"Qayyumabad"
+  },
+  {
+    "name":"Khokrapar"
+  },
+  {
+    "name":"Shah Faisal Colony - Muslimabad Malir City"
+  },
+  {
+    "name":"F.B Area - Block 8"
+  },
+  {
+    "name":"Nanak Wara"
+  },
+  {
+    "name":"Mohammad Ali Society"
+  },
+  {
+    "name":"Manzoor Colony"
+  },
+  {
+    "name":"Dalmia"
+  },
+  {
+    "name":"Defence View - Phase 1"
+  },
+  {
+    "name":"Defence View - Phase 2"
+  },
+  {
+    "name":"KDA Officers Housing Society"
+  },
+  {
+    "name":"Karimabad"
+  },
+  {
+    "name":"Soldier Bazar"
+  },
+  {
+    "name":"Hussainabad"
+  },
+  {
+    "name":"Sharfabad Society"
+  },
+  {
+    "name":"Gharibabad"
+  },
+  {
+    "name":"Sindhi Muslim Cooperative Housing Society"
+  }
+]
+
+export const vendors = [
+  {
+    id:1,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 5,
+    price: "14000",
+    vendorType:"photographer"
+  },
+  {
+    id:2,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 2,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:3,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 3,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:4,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:5,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:6,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:7,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:8,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:9,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:10,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:11,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:12,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:13,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:14,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:15,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:16,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:17,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  { 
+    id:18,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:19,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:20,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:21,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:22,
+    imageUrl:
+      "https://images.unsplash.com/photo-1559508551-44bff1de756b?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=387&q=80",
+    title: "Eyewood Films",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4.4,
+    price: "15000",
+    vendorType:"photographer"
+  },
+  {
+    id:23,
+    imageUrl:
+      "https://images.unsplash.com/photo-1616940844649-535215ae4eb1?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=387&q=80",
+    title: "Ideal Photography",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 4,
+    price: "15000",
+    vendorType:"decorator"
+  },
+  {
+    id:24,
+    imageUrl:
+      "https://images.unsplash.com/photo-1599619351208-3e6c839d6828?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=872&q=80",
+    title: "King Deluxe Room",
+    location: "Ancholi,karachi",
+    reviews: 44,
+    rating: 3.8,
+    price: "15000",
+    vendorType:"decorator"
+  },
+  {
+    id:25,
+    imageUrl:
+      "https://images.unsplash.com/photo-1461092746677-7b4afb1178f6?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=774&q=80",
+    title: "Royal Suite",
+    location: "Ancholi",
+    reviews: 44,
+    rating: 3.5,
+    price: "15000",
+    vendorType:"venue"
+  },
+];
+
