+++ src/styles.css
@@ -1,53 +1,56 @@
 .background-login-signup {
-    background-image: url(./img/login-signup.png);
-    background-repeat: no-repeat;
-    width: 100vw;
-    height: 100vh; 
-    background-size: cover; 
-    background-position: center;
-    position: fixed;
-  }
-
-.font.fontweight-400.fontweight-500.fontweight-600.fontweight-700{
-  font-family:'Public Sans';
- 
+  background-image: url(./img/login-signup.png);
+  background-repeat: no-repeat;
+  width: 100vw;
+  height: 100vh;
+  background-size: cover;
+  background-position: center;
+  position: fixed;
 }
 
-.fontweight-700{
-  font-weight:700;
+.font.fontweight-400.fontweight-500.fontweight-600.fontweight-700 {
+  font-family: "Public Sans";
 }
 
-.fontweight-600{
-  font-weight:600;
+.fontweight-700 {
+  font-weight: 700;
 }
 
-.fontweight-500{
-  font-weight:600;
-}
-.fontweight-400{
-  font-weight:400;
+.fontweight-600 {
+  font-weight: 600;
 }
 
-.btn-outline,.btn-outline:hover, .btn-outline:focus, .btn-outline:active, .btn-outline.active, .open>.dropdown-toggle.btn-outline{
+.fontweight-500 {
+  font-weight: 600;
+}
+.fontweight-400 {
+  font-weight: 400;
+}
+
+.btn-outline,
+.btn-outline:hover,
+.btn-outline:focus,
+.btn-outline:active,
+.btn-outline.active,
+.open > .dropdown-toggle.btn-outline {
   color: black;
   border: 1px solid black;
   background-color: rgb(250, 250, 250);
 }
 
-
-#eye-icon-login-password{
+#eye-icon-login-password {
   position: absolute;
   top: 35%;
   right: 10%;
 }
 
-#eye-icon-signup-password{
+#eye-icon-signup-password {
   position: absolute;
   top: 35%;
   right: 10%;
 }
 
-#eye-icon-signup-cpassword{
+#eye-icon-signup-cpassword {
   position: absolute;
   top: 43%;
   right: 10%;
@@ -58,45 +61,45 @@
   cursor: pointer;
 }
 
-input[type=number]::-webkit-inner-spin-button, 
-input[type=number]::-webkit-outer-spin-button { 
-  -webkit-appearance: none; 
-  margin: 0; 
+input[type="number"]::-webkit-inner-spin-button,
+input[type="number"]::-webkit-outer-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
 }
 
-.item{
+.item {
   object-fit: cover;
   height: 300px;
   max-height: 300px;
-  
 }
 
-.slick-prev:before, .slick-next:before {
+.slick-prev:before,
+.slick-next:before {
   color: black !important;
 }
 
-.card-overlay{
+.card-overlay {
   position: absolute;
-    bottom: 5%;
-    width: 100%;
-    padding-left: 20px;
-    padding-right: 20px;
-    left: 0;
+  bottom: 5%;
+  width: 100%;
+  padding-left: 20px;
+  padding-right: 20px;
+  left: 0;
 }
 
-.star-icon{
+.star-icon {
   color: #ffb400;
 }
 
-.dropdown:hover .dropdown-menu{
+.dropdown:hover .dropdown-menu {
   display: block;
 }
-.dropdown-menu{
+.dropdown-menu {
   margin-top: 0;
 }
 
-.scrollbutton{
-  position: fixed; 
+.scrollbutton {
+  position: fixed;
   right: 1%;
   bottom: 40px;
   font-size: 1.5rem;
@@ -104,6 +107,55 @@ input[type=number]::-webkit-outer-spin-button {
   cursor: pointer;
 }
 
-.pagination-hover{
+.pagination-hover {
   cursor: pointer;
+}
+
+.background-vendor-layout {
+  background-color: #EFF2F6;
+  min-height: 1100px;
+}
+.background-home-body {
+  background-color: #EFF2F6;
+}
+.background-home-footer {
+  background-color: #353434;
+  height: 150px;
+  width: 100%;
+}
+
+.dropdown-filter {
+  width: 15rem;
+  color: gray;
+  height: 2.8rem;
+}
+
+.divider-filter {
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+  border: 0;
+  border-top: 1px solid rgb(0, 0, 0);
+  color: black;
+  width: 100%;
+}
+
+.background-top-specific-vendor{
+    background-color: #D9D9D9;
+    height: 150px;
+    width: 100%;
+  
+}
+
+.background-body-specific-vendor{
+  background-color: #EFF2F6;
+  height: 35rem;
+  width: 100%;
+
+}
+
+.textfield-filter{
+  display: inline;
+  width: 5rem;
+  color: gray;
+  height: 2.8rem;
 }
